requirements.txt
flask
flask-cors
groq
pymongo
python-dotenv
